backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
pyyaml
redis
faiss-cpu
numpy
httpx
python-multipart
torch
onnxruntime
